requirements.txt
Flask==1.1.2
Werkzeug==1.0.1
flask_login==0.5.0
flask_sqlalchemy==2.4.3
flask-bcrypt